Unify window size with new project and new item windows
