+++ code/src/UI/ViewModels/NewProject/MainViewModel.cs
@@ -25,9 +25,6 @@ namespace Microsoft.Templates.UI.ViewModels.NewProject
         public static MainViewModel Current;
         public MainView MainView;
 
-        public double Width { get; }
-        public double Height { get; }
-
         public ProjectSetupViewModel ProjectSetup { get; } = new ProjectSetupViewModel();
 
         public ProjectTemplatesViewModel ProjectTemplates { get; } = new ProjectTemplatesViewModel();
@@ -44,9 +41,6 @@ namespace Microsoft.Templates.UI.ViewModels.NewProject
 
         public MainViewModel() : base()
         {
-            var size = SystemService.Instance.GetMainWindowSize();
-            Width = size.width;
-            Height = size.height;
             Licenses.CollectionChanged += (s, o) => { OnPropertyChanged(nameof(Licenses)); };
             Current = this;
         }
