+++ code/src/UI/Views/NewItem/MainView.xaml
@@ -10,8 +10,8 @@
     PreviewMouseDown="OnPreviewMouseDown"
     Title="{x:Static res:StringRes.Title}"
     Closing="Window_Closing"
-    Width="1277"
-    Height="727"
+    Width="{Binding WizardStatus.Width}"
+    Height="{Binding WizardStatus.Height}"
     MinWidth="800"
     MinHeight="600"
     ResizeMode="CanResize"
