+++ code/src/UI/Views/NewProject/MainView.xaml
@@ -12,8 +12,8 @@
     PreviewMouseDown="OnPreviewMouseDown"
     KeyDown="OnKeyDown"
     Title="{x:Static res:StringRes.Title}"
-    Width="{Binding Width}"
-    Height="{Binding Height}"
+    Width="{Binding WizardStatus.Width}"
+    Height="{Binding WizardStatus.Height}"
     MinWidth="800"
     MinHeight="600"
     ResizeMode="CanResize"
